requirements.txt
coc.py
pandas
matplotlib
python-dotenv
asyncio
Flask